Pass 'alt' text only when confident
It's not evident (based on viewing pins) what the use of alt-text is or
will be. A lot of places, we do ugly things to just stick 'some'
(sometimes) meaningful alt text there. I think the best strategy is to
pass alt text only when it is the 'true' alt text of the image and pass
nothing otherwise until we figure out what meaning it shall have on
Pinterest.